requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
frozenlist==1.5.0
idna==3.10
multidict==6.2.0
propcache==0.3.0
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.0.1
rsa==4.9
Telethon==1.39.0
yarl==1.18.3
